Convert public app parts
