+++ BTCPayServer/Hosting/BTCPayServerServices.cs
@@ -414,7 +414,6 @@ namespace BTCPayServer.Hosting
             services.TryAddScoped<IHttpContextAccessor, HttpContextAccessor>();
             services.AddTransient<BitpayAccessTokenController>();
             services.AddTransient<UIInvoiceController>();
-            services.AddTransient<UIAppsPublicController>();
             services.AddTransient<UIPaymentRequestController>();
             // Add application services.
             services.AddSingleton<EmailSenderFactory>();
