+++ BTCPayServer/Plugins/Crowdfund/Controllers/UICrowdfundController.cs
@@ -1,41 +1,211 @@
 using System;
+using System.Collections.Generic;
 using System.Linq;
+using System.Threading;
 using System.Threading.Tasks;
 using BTCPayServer.Abstractions.Constants;
+using BTCPayServer.Abstractions.Extensions;
 using BTCPayServer.Client;
 using BTCPayServer.Controllers;
 using BTCPayServer.Data;
+using BTCPayServer.Filters;
+using BTCPayServer.Models;
 using BTCPayServer.Plugins.Crowdfund.Models;
+using BTCPayServer.Plugins.PointOfSale.Models;
 using BTCPayServer.Services.Apps;
 using BTCPayServer.Services.Rates;
 using BTCPayServer.Services.Stores;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Cors;
+using Microsoft.AspNetCore.Http.Extensions;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
+using NBitpayClient;
+using NicolasDorier.RateLimits;
 
 namespace BTCPayServer.Plugins.Crowdfund.Controllers
 {
-    [Authorize(Policy = Policies.CanModifyStoreSettings, AuthenticationSchemes = AuthenticationSchemes.Cookie)]
     [AutoValidateAntiforgeryToken]
     [Route("apps")]
     public class UICrowdfundController : Controller
     {
         public UICrowdfundController(
-            EventAggregator eventAggregator,
+            AppService appService,
             CurrencyNameTable currencies,
+            EventAggregator eventAggregator,
             StoreRepository storeRepository,
-            AppService appService)
+            UIInvoiceController invoiceController,
+            UserManager<ApplicationUser> userManager)
         {
-            _eventAggregator = eventAggregator;
             _currencies = currencies;
-            _storeRepository = storeRepository;
             _appService = appService;
+            _userManager = userManager;
+            _storeRepository = storeRepository;
+            _eventAggregator = eventAggregator;
+            _invoiceController = invoiceController;
         }
 
         private readonly EventAggregator _eventAggregator;
         private readonly CurrencyNameTable _currencies;
         private readonly StoreRepository _storeRepository;
         private readonly AppService _appService;
+        private readonly UIInvoiceController _invoiceController;
+        private readonly UserManager<ApplicationUser> _userManager;
+        
+        [HttpGet("/")]
+        [HttpGet("/apps/{appId}/crowdfund")]
+        [XFrameOptions(XFrameOptionsAttribute.XFrameOptions.AllowAll)]
+        [DomainMappingConstraint(AppType.Crowdfund)]
+        public async Task<IActionResult> ViewCrowdfund(string appId, string statusMessage)
+        {
+            var app = await _appService.GetApp(appId, AppType.Crowdfund, true);
 
+            if (app == null)
+                return NotFound();
+            var settings = app.GetSettings<CrowdfundSettings>();
+
+            var isAdmin = await _appService.GetAppDataIfOwner(GetUserId(), appId, AppType.Crowdfund) != null;
+
+            var hasEnoughSettingsToLoad = !string.IsNullOrEmpty(settings.TargetCurrency);
+            if (!hasEnoughSettingsToLoad)
+            {
+                if (!isAdmin)
+                    return NotFound();
+
+                return NotFound("A Target Currency must be set for this app in order to be loadable.");
+            }
+            var appInfo = await GetAppInfo(appId);
+
+            if (settings.Enabled)
+                return View("Crowdfund/Public/ViewCrowdfund", appInfo);
+            if (!isAdmin)
+                return NotFound();
+
+            return View("Crowdfund/Public/ViewCrowdfund", appInfo);
+        }
+
+        [HttpPost("/")]
+        [HttpPost("/apps/{appId}/crowdfund")]
+        [XFrameOptions(XFrameOptionsAttribute.XFrameOptions.AllowAll)]
+        [IgnoreAntiforgeryToken]
+        [EnableCors(CorsPolicies.All)]
+        [DomainMappingConstraint(AppType.Crowdfund)]
+        [RateLimitsFilter(ZoneLimits.PublicInvoices, Scope = RateLimitsScope.RemoteAddress)]
+        public async Task<IActionResult> ContributeToCrowdfund(string appId, ContributeToCrowdfund request, CancellationToken cancellationToken)
+        {
+
+            var app = await _appService.GetApp(appId, AppType.Crowdfund, true);
+
+            if (app == null)
+                return NotFound();
+            var settings = app.GetSettings<CrowdfundSettings>();
+
+            var isAdmin = await _appService.GetAppDataIfOwner(GetUserId(), appId, AppType.Crowdfund) != null;
+
+            if (!settings.Enabled && !isAdmin)
+            {
+                return NotFound("Crowdfund is not currently active");
+            }
+
+            var info = await GetAppInfo(appId);
+            if (!isAdmin &&
+                ((settings.StartDate.HasValue && DateTime.UtcNow < settings.StartDate) ||
+                 (settings.EndDate.HasValue && DateTime.UtcNow > settings.EndDate) ||
+                 (settings.EnforceTargetAmount &&
+                  (info.Info.PendingProgressPercentage.GetValueOrDefault(0) +
+                   info.Info.ProgressPercentage.GetValueOrDefault(0)) >= 100)))
+            {
+                return NotFound("Crowdfund is not currently active");
+            }
+
+            var store = await _appService.GetStore(app);
+            var title = settings.Title;
+            decimal? price = request.Amount;
+            Dictionary<string, InvoiceSupportedTransactionCurrency> paymentMethods = null;
+            ViewPointOfSaleViewModel.Item choice = null;
+            if (!string.IsNullOrEmpty(request.ChoiceKey))
+            {
+                var choices = _appService.GetPOSItems(settings.PerksTemplate, settings.TargetCurrency);
+                choice = choices.FirstOrDefault(c => c.Id == request.ChoiceKey);
+                if (choice == null)
+                    return NotFound("Incorrect option provided");
+                title = choice.Title;
+
+                if (choice.Price.Type == ViewPointOfSaleViewModel.Item.ItemPrice.ItemPriceType.Topup)
+                {
+                    price = null;
+                }
+                else
+                {
+                    price = choice.Price.Value;
+                    if (request.Amount > price)
+                        price = request.Amount;
+                }
+                if (choice.Inventory.HasValue)
+                {
+                    if (choice.Inventory <= 0)
+                    {
+                        return NotFound("Option was out of stock");
+                    }
+                }
+                if (choice?.PaymentMethods?.Any() is true)
+                {
+                    paymentMethods = choice?.PaymentMethods.ToDictionary(s => s,
+                        s => new InvoiceSupportedTransactionCurrency() { Enabled = true });
+                }
+            }
+            else
+            {
+                if (request.Amount < 0)
+                {
+                    return NotFound("Please provide an amount greater than 0");
+                }
+
+                price = request.Amount;
+            }
+
+            if (!isAdmin && (settings.EnforceTargetAmount && info.TargetAmount.HasValue && price >
+                             (info.TargetAmount - (info.Info.CurrentAmount + info.Info.CurrentPendingAmount))))
+            {
+                return NotFound("Contribution Amount is more than is currently allowed.");
+            }
+
+            try
+            {
+                var invoice = await _invoiceController.CreateInvoiceCore(new BitpayCreateInvoiceRequest()
+                    {
+                        OrderId = AppService.GetAppOrderId(app),
+                        Currency = settings.TargetCurrency,
+                        ItemCode = request.ChoiceKey ?? string.Empty,
+                        ItemDesc = title,
+                        BuyerEmail = request.Email,
+                        Price = price,
+                        NotificationURL = settings.NotificationUrl,
+                        FullNotifications = true,
+                        ExtendedNotifications = true,
+                        SupportedTransactionCurrencies = paymentMethods,
+                        RedirectURL = request.RedirectUrl ?? Request.GetDisplayUrl(),
+                    }, store, HttpContext.Request.GetAbsoluteRoot(),
+                    new List<string>() {AppService.GetAppInternalTag(appId)},
+                    cancellationToken, (entity) =>
+                    {
+                        entity.Metadata.OrderUrl = Request.GetDisplayUrl();
+                    });
+
+                if (request.RedirectToCheckout)
+                {
+                    return RedirectToAction(nameof(UIInvoiceController.Checkout), "UIInvoice",
+                        new {invoiceId = invoice.Data.Id});
+                }
+
+                return Ok(invoice.Data.Id);
+            }
+            catch (BitpayHttpException e)
+            {
+                return BadRequest(e.Message);
+            }
+        }
+        
         [Authorize(Policy = Policies.CanModifyStoreSettings, AuthenticationSchemes = AuthenticationSchemes.Cookie)]
         [HttpGet("{appId}/settings/crowdfund")]
         public async Task<IActionResult> UpdateCrowdfund(string appId)
@@ -85,6 +255,7 @@ namespace BTCPayServer.Plugins.Crowdfund.Controllers
             return View("Crowdfund/UpdateCrowdfund", vm);
         }
 
+        [Authorize(Policy = Policies.CanModifyStoreSettings, AuthenticationSchemes = AuthenticationSchemes.Cookie)]
         [HttpPost("{appId}/settings/crowdfund")]
         public async Task<IActionResult> UpdateCrowdfund(string appId, UpdateCrowdfundViewModel vm, string command)
         {
@@ -199,7 +370,7 @@ namespace BTCPayServer.Plugins.Crowdfund.Controllers
             return RedirectToAction(nameof(UpdateCrowdfund), new { appId });
         }
 
-        async Task<string> GetStoreDefaultCurrentIfEmpty(string storeId, string currency)
+        private async Task<string> GetStoreDefaultCurrentIfEmpty(string storeId, string currency)
         {
             if (string.IsNullOrWhiteSpace(currency))
             {
@@ -209,5 +380,15 @@ namespace BTCPayServer.Plugins.Crowdfund.Controllers
         }
         
         private AppData GetCurrentApp() => HttpContext.GetAppData();
+
+        private string GetUserId() => _userManager.GetUserId(User);
+        
+        private async Task<ViewCrowdfundViewModel> GetAppInfo(string appId)
+        {
+            var info = (ViewCrowdfundViewModel)await _appService.GetAppInfo(appId);
+            info.HubPath = AppHub.GetHubPath(Request);
+            info.SimpleDisplay = Request.Query.ContainsKey("simple");
+            return info;
+        }
     }
 }
