+++ BTCPayServer/Plugins/PointOfSale/Controllers/UIPointOfSaleController.cs
@@ -1,42 +1,252 @@
 using System;
+using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
 using System.Text;
 using System.Text.Encodings.Web;
 using System.Text.RegularExpressions;
+using System.Threading;
 using System.Threading.Tasks;
 using BTCPayServer.Abstractions.Constants;
 using BTCPayServer.Abstractions.Extensions;
+using BTCPayServer.Abstractions.Models;
 using BTCPayServer.Client;
+using BTCPayServer.Controllers;
 using BTCPayServer.Data;
+using BTCPayServer.Filters;
+using BTCPayServer.ModelBinders;
+using BTCPayServer.Models;
 using BTCPayServer.Plugins.PointOfSale.Models;
 using BTCPayServer.Services.Apps;
 using BTCPayServer.Services.Rates;
 using BTCPayServer.Services.Stores;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Cors;
+using Microsoft.AspNetCore.Http.Extensions;
 using Microsoft.AspNetCore.Mvc;
+using NBitpayClient;
+using NicolasDorier.RateLimits;
 
 namespace BTCPayServer.Plugins.PointOfSale.Controllers
 {
-    [Authorize(Policy = Policies.CanModifyStoreSettings, AuthenticationSchemes = AuthenticationSchemes.Cookie)]
     [AutoValidateAntiforgeryToken]
     [Route("apps")]
     public class UIPointOfSaleController : Controller
     {
         public UIPointOfSaleController(
+            AppService appService,
             CurrencyNameTable currencies,
             StoreRepository storeRepository,
-            AppService appService)
+            UIInvoiceController invoiceController)
         {
             _currencies = currencies;
-            _storeRepository = storeRepository;
             _appService = appService;
+            _storeRepository = storeRepository;
+            _invoiceController = invoiceController;
         }
 
         private readonly CurrencyNameTable _currencies;
         private readonly StoreRepository _storeRepository;
         private readonly AppService _appService;
+        private readonly UIInvoiceController _invoiceController;
         
+        [HttpGet("/")]
+        [HttpGet("/apps/{appId}/pos/{viewType?}")]
+        [XFrameOptions(XFrameOptionsAttribute.XFrameOptions.AllowAll)]
+        [DomainMappingConstraint(AppType.PointOfSale)]
+        public async Task<IActionResult> ViewPointOfSale(string appId, PosViewType? viewType = null)
+        {
+            var app = await _appService.GetApp(appId, AppType.PointOfSale);
+            if (app == null)
+                return NotFound();
+            var settings = app.GetSettings<PointOfSaleSettings>();
+            var numberFormatInfo = _appService.Currencies.GetNumberFormatInfo(settings.Currency) ?? 
+                                   _appService.Currencies.GetNumberFormatInfo("USD");
+            double step = Math.Pow(10, -numberFormatInfo.CurrencyDecimalDigits);
+            viewType ??= settings.EnableShoppingCart ? PosViewType.Cart : settings.DefaultView;
+            var store = await _appService.GetStore(app);
+            var storeBlob = store.GetStoreBlob();
+
+            return View($"PointOfSale/Public/{viewType}", new ViewPointOfSaleViewModel
+            {
+                Title = settings.Title,
+                Step = step.ToString(CultureInfo.InvariantCulture),
+                ViewType = (PosViewType)viewType,
+                ShowCustomAmount = settings.ShowCustomAmount,
+                ShowDiscount = settings.ShowDiscount,
+                EnableTips = settings.EnableTips,
+                CurrencyCode = settings.Currency,
+                CurrencySymbol = numberFormatInfo.CurrencySymbol,
+                CurrencyInfo = new ViewPointOfSaleViewModel.CurrencyInfoData
+                {
+                    CurrencySymbol = string.IsNullOrEmpty(numberFormatInfo.CurrencySymbol) ? settings.Currency : numberFormatInfo.CurrencySymbol,
+                    Divisibility = numberFormatInfo.CurrencyDecimalDigits,
+                    DecimalSeparator = numberFormatInfo.CurrencyDecimalSeparator,
+                    ThousandSeparator = numberFormatInfo.NumberGroupSeparator,
+                    Prefixed = new[] { 0, 2 }.Contains(numberFormatInfo.CurrencyPositivePattern),
+                    SymbolSpace = new[] { 2, 3 }.Contains(numberFormatInfo.CurrencyPositivePattern)
+                },
+                Items = _appService.GetPOSItems(settings.Template, settings.Currency),
+                ButtonText = settings.ButtonText,
+                CustomButtonText = settings.CustomButtonText,
+                CustomTipText = settings.CustomTipText,
+                CustomTipPercentages = settings.CustomTipPercentages,
+                CustomCSSLink = settings.CustomCSSLink,
+                CustomLogoLink = storeBlob.CustomLogo,
+                AppId = appId,
+                StoreId = store.Id,
+                Description = settings.Description,
+                EmbeddedCSS = settings.EmbeddedCSS,
+                RequiresRefundEmail = settings.RequiresRefundEmail
+            });
+        }
+
+        [HttpPost("/")]
+        [HttpPost("/apps/{appId}/pos/{viewType?}")]
+        [XFrameOptions(XFrameOptionsAttribute.XFrameOptions.AllowAll)]
+        [IgnoreAntiforgeryToken]
+        [EnableCors(CorsPolicies.All)]
+        [DomainMappingConstraint(AppType.PointOfSale)]
+        [RateLimitsFilter(ZoneLimits.PublicInvoices, Scope = RateLimitsScope.RemoteAddress)]
+        public async Task<IActionResult> ViewPointOfSale(string appId,
+                                                        PosViewType viewType,
+                                                        [ModelBinder(typeof(InvariantDecimalModelBinder))] decimal? amount,
+                                                        string email,
+                                                        string orderId,
+                                                        string notificationUrl,
+                                                        string redirectUrl,
+                                                        string choiceKey,
+                                                        string posData = null,
+                                                        RequiresRefundEmail requiresRefundEmail = RequiresRefundEmail.InheritFromStore,
+                                                        CancellationToken cancellationToken = default)
+        {
+            var app = await _appService.GetApp(appId, AppType.PointOfSale);
+            if (string.IsNullOrEmpty(choiceKey) && amount <= 0)
+            {
+                return RedirectToAction(nameof(ViewPointOfSale), new { appId });
+            }
+            if (app == null)
+                return NotFound();
+            var settings = app.GetSettings<PointOfSaleSettings>();
+            settings.DefaultView = settings.EnableShoppingCart ? PosViewType.Cart : settings.DefaultView;
+            if (string.IsNullOrEmpty(choiceKey) && !settings.ShowCustomAmount && settings.DefaultView != PosViewType.Cart)
+            {
+                return RedirectToAction(nameof(ViewPointOfSale), new { appId, viewType });
+            }
+            string title = null;
+            decimal? price = null;
+            Dictionary<string, InvoiceSupportedTransactionCurrency> paymentMethods = null;
+            ViewPointOfSaleViewModel.Item choice = null;
+            if (!string.IsNullOrEmpty(choiceKey))
+            {
+                var choices = _appService.GetPOSItems(settings.Template, settings.Currency);
+                choice = choices.FirstOrDefault(c => c.Id == choiceKey);
+                if (choice == null)
+                    return NotFound();
+                title = choice.Title;
+                if (choice.Price.Type == ViewPointOfSaleViewModel.Item.ItemPrice.ItemPriceType.Topup)
+                {
+                    price = null;
+                }
+                else
+                {
+                    price = choice.Price.Value;
+                    if (amount > price)
+                        price = amount;
+                }
+
+                if (choice.Inventory.HasValue)
+                {
+                    if (choice.Inventory <= 0)
+                    {
+                        return RedirectToAction(nameof(ViewPointOfSale), new { appId = appId });
+                    }
+                }
+
+                if (choice?.PaymentMethods?.Any() is true)
+                {
+                    paymentMethods = choice?.PaymentMethods.ToDictionary(s => s,
+                        s => new InvoiceSupportedTransactionCurrency() { Enabled = true });
+                }
+            }
+            else
+            {
+                if (!settings.ShowCustomAmount && settings.DefaultView != PosViewType.Cart)
+                    return NotFound();
+                price = amount;
+                title = settings.Title;
+
+                //if cart IS enabled and we detect posdata that matches the cart system's, check inventory for the items
+                if (!string.IsNullOrEmpty(posData) &&
+                    settings.DefaultView == PosViewType.Cart &&
+                    AppService.TryParsePosCartItems(posData, out var cartItems))
+                {
+
+                    var choices = _appService.GetPOSItems(settings.Template, settings.Currency);
+                    foreach (var cartItem in cartItems)
+                    {
+                        var itemChoice = choices.FirstOrDefault(c => c.Id == cartItem.Key);
+                        if (itemChoice == null)
+                            return NotFound();
+
+                        if (itemChoice.Inventory.HasValue)
+                        {
+                            switch (itemChoice.Inventory)
+                            {
+                                case int i when i <= 0:
+                                    return RedirectToAction(nameof(ViewPointOfSale), new { appId });
+                                case int inventory when inventory < cartItem.Value:
+                                    return RedirectToAction(nameof(ViewPointOfSale), new { appId });
+                            }
+                        }
+                    }
+                }
+            }
+            var store = await _appService.GetStore(app);
+            try
+            {
+                var invoice = await _invoiceController.CreateInvoiceCore(new BitpayCreateInvoiceRequest()
+                {
+                    ItemCode = choice?.Id,
+                    ItemDesc = title,
+                    Currency = settings.Currency,
+                    Price = price,
+                    BuyerEmail = email,
+                    OrderId = orderId ?? AppService.GetAppOrderId(app),
+                    NotificationURL =
+                            string.IsNullOrEmpty(notificationUrl) ? settings.NotificationUrl : notificationUrl,
+                    RedirectURL =  !string.IsNullOrEmpty(redirectUrl) ? redirectUrl
+                        : !string.IsNullOrEmpty(settings.RedirectUrl) ? settings.RedirectUrl
+                        : Request.GetDisplayUrl(),
+                    FullNotifications = true,
+                    ExtendedNotifications = true,
+                    PosData = string.IsNullOrEmpty(posData) ? null : posData,
+                    RedirectAutomatically = settings.RedirectAutomatically,
+                    SupportedTransactionCurrencies = paymentMethods,
+                    RequiresRefundEmail = requiresRefundEmail == RequiresRefundEmail.InheritFromStore
+                        ? store.GetStoreBlob().RequiresRefundEmail
+                        : requiresRefundEmail == RequiresRefundEmail.On,
+                }, store, HttpContext.Request.GetAbsoluteRoot(),
+                    new List<string>() { AppService.GetAppInternalTag(appId) },
+                    cancellationToken, (entity) =>
+                    {
+                        entity.Metadata.OrderUrl = Request.GetDisplayUrl();
+                    } );
+                return RedirectToAction(nameof(UIInvoiceController.Checkout), "UIInvoice", new { invoiceId = invoice.Data.Id });
+            }
+            catch (BitpayHttpException e)
+            {
+                TempData.SetStatusMessageModel(new StatusMessageModel
+                {
+                    Html = e.Message.Replace("\n", "<br />", StringComparison.OrdinalIgnoreCase),
+                    Severity = StatusMessageModel.StatusSeverity.Error,
+                    AllowDismiss = true
+                });
+                return RedirectToAction(nameof(ViewPointOfSale), new { appId = appId });
+            }
+        }
+        
+        [Authorize(Policy = Policies.CanModifyStoreSettings, AuthenticationSchemes = AuthenticationSchemes.Cookie)]
         [HttpGet("{appId}/settings/pos")]
         public async Task<IActionResult> UpdatePointOfSale(string appId)
         {
@@ -113,6 +323,7 @@ namespace BTCPayServer.Plugins.PointOfSale.Controllers
             return View("PointOfSale/UpdatePointOfSale", vm);
         }
 
+        [Authorize(Policy = Policies.CanModifyStoreSettings, AuthenticationSchemes = AuthenticationSchemes.Cookie)]
         [HttpPost("{appId}/settings/pos")]
         public async Task<IActionResult> UpdatePointOfSale(string appId, UpdatePointOfSaleViewModel vm)
         {
@@ -180,7 +391,7 @@ namespace BTCPayServer.Plugins.PointOfSale.Controllers
             return list.Split(separator, StringSplitOptions.RemoveEmptyEntries).Select(int.Parse).ToArray();
         }
 
-        async Task<string> GetStoreDefaultCurrentIfEmpty(string storeId, string currency)
+        private async Task<string> GetStoreDefaultCurrentIfEmpty(string storeId, string currency)
         {
             if (string.IsNullOrWhiteSpace(currency))
             {
