+++ BTCPayServer/Views/Shared/Crowdfund/Public/ViewCrowdfund.cshtml
@@ -252,7 +252,7 @@
             </div>
             <div class="col-md-4 col-sm-12">
                 <partial
-                    name="/Views/UIAppsPublic/Crowdfund/ContributeForm.cshtml"
+                    name="Crowdfund/Public/ContributeForm"
                     model="@(new ContributeToCrowdfund { ViewCrowdfundViewModel = Model, RedirectToCheckout = true })">
                 </partial>
             </div>
