+++ BTCPayServer/Views/Shared/Crowdfund/UpdateCrowdfund.cshtml
@@ -31,7 +31,7 @@
             <button type="submit" class="btn btn-primary order-sm-1" id="SaveSettings">Save</button>   
             @if (Model.ModelWithMinimumData)
             {
-                <a class="btn btn-secondary" asp-action="ViewCrowdfund" asp-controller="UIAppsPublic" asp-route-appId="@Model.AppId" id="ViewApp" target="_blank">View</a>
+                <a class="btn btn-secondary" asp-action="ViewCrowdfund" asp-route-appId="@Model.AppId" id="ViewApp" target="_blank">View</a>
             }
         </div>
     </div>
