+++ BTCPayServer/Views/Shared/PointOfSale/Public/MinimalLight.cshtml
@@ -7,7 +7,7 @@
         }
     </div>
     <div class="py-5 px-3">
-        <form method="post" asp-controller="UIAppsPublic" asp-action="ViewPointOfSale" asp-route-appId="@Model.AppId" asp-antiforgery="false" data-buy>
+        <form method="post" asp-action="ViewPointOfSale" asp-route-appId="@Model.AppId" asp-antiforgery="false" data-buy>
             <div class="input-group">
                 <span class="input-group-text">@Model.CurrencySymbol</span>
                 <input class="form-control" type="number" step="@Model.Step" name="amount" placeholder="Amount">
