+++ BTCPayServer/Views/Shared/PointOfSale/Public/Print.cshtml
@@ -1,4 +1,3 @@
-@using BTCPayServer.Models.AppViewModels
 @using BTCPayServer.Payments.Lightning
 @using BTCPayServer.Plugins.PointOfSale.Models
 @using BTCPayServer.Services.Stores
@@ -20,7 +19,7 @@
 
 @{
     var store = await StoreRepository.FindStore(Model.StoreId);
-    Layout = "_LayoutPos";
+    Layout = "PointOfSale/Public/_LayoutPos";
     Context.Request.Query.TryGetValue("cryptocode", out var cryptoCodeValues);
     var cryptoCode = cryptoCodeValues.FirstOrDefault() ?? "BTC";
     var supported = store.GetSupportedPaymentMethods(BTCPayNetworkProvider).OfType<LNURLPaySupportedPaymentMethod>().OrderBy(method => method.CryptoCode == cryptoCode).FirstOrDefault();
