+++ BTCPayServer/Views/UIApps/ListApps.cshtml
@@ -103,10 +103,6 @@
                                 <a asp-action="@app.UpdateAction" asp-controller="UIApps" asp-route-appId="@app.Id" asp-route-storeId="@app.StoreId">Settings</a>
                                 <span> - </span>
                             }
-                            <a asp-action="@app.ViewAction" asp-controller="UIAppsPublic" asp-route-appId="@app.Id">View</a>
-                            <a asp-action="@app.ViewAction" asp-controller="UIAppsPublic" asp-route-appId="@app.Id" target="_blank"
-                               title="View in New Window"><span class="fa fa-external-link"></span></a>
-                            <span> - </span>
                             <a asp-action="DeleteApp" asp-route-appId="@app.Id" data-bs-toggle="modal" data-bs-target="#ConfirmModal" data-description="The app <strong>@app.AppName</strong> and its settings will be permanently deleted from your store <strong>@app.StoreName</strong>." data-confirm-input="DELETE">Delete</a>
                         </td>
                     </tr>
