--- BTCPayServer/Views/UIAppsPublic/PointOfSale/Light.cshtml
@@ -1,18 +1,0 @@
-@model BTCPayServer.Plugins.PointOfSale.Models.ViewPointOfSaleViewModel
-@{
-    Layout = "_LayoutPos";
-}
-
-<partial name="_StatusMessage" />
-
-@if (Context.Request.Query.ContainsKey("simple"))
-{
-    <partial name="/Views/UIAppsPublic/PointOfSale/MinimalLight.cshtml" model="Model" />
-}
-else
-{
-    <noscript>
-        <partial name="/Views/UIAppsPublic/PointOfSale/MinimalLight.cshtml" model="Model" />
-    </noscript>
-    <partial name="/Views/UIAppsPublic/PointOfSale/VueLight.cshtml" model="Model" />
-}
